Create code
This function takes two parameters startDate and endDate which specify the date range for which Pedometer data should be fetched. It constructs a URL for the API endpoint using the start and end date parameters, and then makes a request to that URL using the fetch() function.

If the response from the API is successful, the function extracts the steps data from the JSON response and returns it. If there is an error, it logs an error message to the console.

Note that this is just an example and the exact implementation of this function would depend on the specifics of the Pedometer API being used.